Parameter Table Helpers

Helper functions to manipulate parameter tables.

Usage

combine_partables(object_list, drop_duplicated = TRUE)

# S3 method for class 'eq_partables'
print(
  x,
  max_models = NULL,
  names_to_use = c("default", "long"),
  wrap_long_names = TRUE,
  readable_long_names = TRUE,
  ...
)

# S3 method for class 'eq_partables'
c(..., drop_duplicated = TRUE)

# S3 method for class 'eq_partables'
duplicated(x, incomparables = FALSE, ...)

# S3 method for class 'eq_partables'
unique(x, incomparables = FALSE, ...)

setdiff_eq_partables(x, y)

union_eq_partables(x, y)

intersect_eq_partables(x, y)

setequal_eq_partables(x, y)

is_element_eq_partables(el, set)

match_eq_partables(x, table, nomatch = NA_integer_)

x %pt_in% table

x %pt_notin% table

eq_partables(...)

as_eq_partables(sem_out = NULL, model_name = "original")

is_partable(object, colchk = c("id", "lhs", "op", "rhs"))

is_partables(object, colchk = c("id", "lhs", "op", "rhs"))

Arguments

object_list

A list of objects of the class partables or eq_partables.

drop_duplicated

Logical. Whether duplicated models will be removed.

x

An eq_partables object.

max_models

The maximum number of models to print. If NULL, all models will be printed.

names_to_use

If "default", the names in x, which may not be descriptive, will be used. If "long", the names from modelbpp::gen_models() will be used if available. They can be very long, but describes the changes leading to a model.

wrap_long_names

If TRUE, long names will be wrapped when printed. Used only when names_to_use is "long".

readable_long_names

If TRUE, the long names will be modified to make them more readable.`

...

For the c-method of eq_partables object, these are eq_partables objects to be combined. For the print-method of eq_partables objects, these arguments are not used. For eq_partables(), it should be lavaan outputs or lavaan parameter tables.

incomparables

Not used.

y

An eq_partables object.

el

A parameter table.

set

A list of parameter tables.

table

A list of parameter tables.

nomatch

The same argument from match().

sem_out

A lavaan object or a lavaan parameter table. Can be NULL.

model_name

The name of the model in the output.

object

The object to be checked whether it is a parameter table.

colchk

The columns to be checked.

Value

The function combine_partables() always returns an object of the class eq_partables.

The print-method of eq_partables return x invisibly. It is called for its side-effet.

The c-method of eq_partables returns a list of the class eq_partables.

The duplicated-method of eq_partables returns a logical vector to indicate which models, if any, are identical to other models earlier in the list.

The unique-method of eq_partables returns an eq_partables object.

The function setdiff_eq_partables() returns a list of parameter tables, of the same class of x, with models present in y removed.

The function union_eq_partables() returns a list of parameter tables, of the class eq_partables.

The function intersect_eq_partables() returns a list of parameter tables, of the class eq_partables, common in both x and y.

The function setequal_eq_partables() returns TRUE or FALSE, based on the results of get_digest_partables() applied to x and y.

The function is_element_eq_partables() returns TRUE or FALSE, based on the results of is.element() applied to the hash values from get_digest() and get_digest_partables().

The function match_eq_partables() returns the results of match() applied to the hash values generated by get_digest().

%pt_in% returns a logical vector, the output of %in% applied to the has values of x and table.

%pt_notin% returns a logical vector, the output of %notin% applied to the has values of x and table.

The function eq_partables() returns an eq_partables object created from the one or more lavaan outputs or lavaan parameter tables.

The function as_eq_partables() returns a one-element eq_partables object if sem_out is a lavaan output or a lavaan parameter table. It returns a zero-length eq_partables object otherwise.

The function is_partable() return either TRUE or FALSE. It is TRUE is the two conditions mentioned in Details are met.

The function is_partables() return either TRUE or FALSE. It is TRUE only if is_partable() returns TRUE for all its elements.

Details

The function combine_partables() combine a list of partables objects or eq_partables objects to one single object of the same type.

The print-method of eq_partables object handles a zero-length list. If not of zero-length, the print-method for partables will be used.

The c-method of eq_partables object combine eq_partables elements to one single eq_partables elements.

The duplicated-method of eq_partables checks whether any models are identical. If yes, the duplicated models, except for the first one, will be denoted as duplicated.

The unique-method of eq_partables returns an eq_partables object with duplicated models, if any, removed.

The function setdiff_eq_partables() removes from x models that are also in y.

The function union_eq_partables() combines the models in x and y, with duplicated models removed.

The function intersect_eq_partables() find the models common in x and y.

The function setequal_eq_partables() check whether x and y has the same set of models. Orders are ignored.

The function is_element_eq_partables() checks whether el is one of the models in set.

The function match_eq_partables() is similar to match(), but check matches based on the results of get_digest() applied to the parameter tables.

%pt_in% is similar to %in%, but works on parameter tables using match_eq_partables().

%pt_notin% is similar to %notin%, but works on parameter tables using match_eq_partables().

The function eq_partables() creates an eq_partables object from lavaan parameter tables or lavaan outputs.

The function as_eq_partables() is not a usual as function. It works only on a lavaan output or lavaan parameter table. It converts sem_out to a one-element list of the class eq_partables, with the parameter table as the element and the lavaan output, if sem_out is a lavaan output, in the attribute "fit". If sem_out is not a lavaan object nor a lavaan parameter table, a zero-length eq_partables object will be returned.

The function is_partable() check whether an object is probably a parameter table. It checks whether (a) the object is a data.frame-like object (by is.data.frame()) and (b) the columns in colchk exist. If both conditions are met, then the object is considered a parameter table.

The function is_partables() check whether a list is likely a list of parameter tables. It simply call is_partable() on all the elements.

Examples

library(lavaan)

# Model 1

mod1 <-
"
fx =~ x1 + x2 + x3
fm =~ m1 + m2 + m3
fy =~ y1 + y2 + y3
fm ~ fx
fy ~ fm + fx
"
fit1 <- sem(
          model = mod1,
          data = data_test_3_factor_3_item
        )
fit1_1_more <- drop_k(fit1)
#> Error in eval(x): object 'mod1' not found
fit1_1_more_1_less <- lapply(
  fit1_1_more,
  add_k
)
#> Error: object 'fit1_1_more' not found

partables1 <- combine_partables(fit1_1_more_1_less)
#> Error: object 'fit1_1_more_1_less' not found
partables1
#> Error: object 'partables1' not found